moving test to its own file
